requirements.txt
openai_whisper==20230314
youtube_dl==2021.12.17
